pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-prompt-injection-benchmark"
version = "0.1.0"
description = "A benchmark for evaluating LLM agent robustness against prompt injection attacks"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Benchmark Team"}
]
keywords = ["llm", "security", "prompt-injection", "benchmark", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "gym>=0.26.2",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[tool.setuptools]
packages = ["envs", "llm", "attacks", "defenses", "runner", "experiments", "analysis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
